src/requirements.txt
pyserial
XlsxWriter
matplotlib
